combinedroot: Use cpio -u
This makes cpio overwrite the busybox binary so we have a sane
dynamically linked one in recovery.  Without this, we have a
badly behaving statically linked one.

Among other things, backuptool fails due to pipes not flushing properly,
likely similar to a similar issue as
http://lists.busybox.net/pipermail/busybox/2006-October/059144.html

This does not appear to affect CWM-based recoveries as:
1)  It is linked slightly differently than static busybox
2)  In CWM-based recoveries, CWM itself has busybox built in, so a bad
/sbin/busybox isn't used

Change-Id: I2d8d201f8e4f4adc9df407b74e8e877507c0abd6